[build] Add mxe commit sha to AndroidMxeInstallPrefix
This ensures that we can revert mxe to an earlier version.

Background:

The current "do we need to rebuild <dependency>" is mostly "forward-only".
It works properly if you bump forward in time, but if you go _backwards_,
it won't rebuild because it's timestamp based.

For example: if we bump MXE, it'll rebuild the (cached!) MXE on the build server
but if we find that MXE is broken and we need to revert... we can't because
what we're reverting to is older than the bump!